Better recommendation removal, auto loading broken.
Now by looking for "shelf annotation shelf-title-annotation" class, all recommendations, popular channels and similar are removed with a much shorter code.
More importantly, auto-loading is deliberately broken in order to prevent YouTube from loading more recommendations and popular channels. It might be possible to use "data-scrolldetect-callback" to listen for changes in the page and remove the recommendations instead.